Approved by Brandon Schaefer, PS Jenkins bot.

(bzr r3343)

+++ UnityCore/Model.h
@@ -63,15 +63,15 @@ public:
   nux::Property<std::string> swarm_name;
 
   nux::ROProperty<std::size_t> count;
-  nux::ROProperty<unsigned long long> seqnum;
+  nux::ROProperty<uint64_t> seqnum;
   nux::ROProperty<glib::Object<DeeModel>> model;
 
   sigc::signal<void, RowAdaptor&> row_added;
   sigc::signal<void, RowAdaptor&> row_changed;
   sigc::signal<void, RowAdaptor&> row_removed;
 
-  sigc::signal<void, unsigned long long, unsigned long long> begin_transaction;
-  sigc::signal<void, unsigned long long, unsigned long long> end_transaction;
+  sigc::signal<void, uint64_t, uint64_t> begin_transaction;
+  sigc::signal<void, uint64_t, uint64_t> end_transaction;
 
   typedef std::function<DeeModelTag*(glib::Object<DeeModel> const& model)> GetDeeTagFunc;
 
@@ -87,7 +87,7 @@ private:
   void OnTransactionEnd(DeeModel* model, guint64 begin_seq, guint64 end_seq);
   void OnSwarmNameChanged(std::string const& swarm_name);
   std::size_t get_count() const;
-  unsigned long long get_seqnum() const;
+  uint64_t get_seqnum() const;
   glib::Object<DeeModel> get_model() const;
 
 private:
